:sparkles: added routing with gameId
