+++ apps/web/src/components/functional/GameScene.tsx
@@ -1,4 +1,4 @@
-import type { PlayersChoices } from '../../routes/shifumi.lazy'
+import type { PlayersChoices } from '../../routes/shifumi.$gameId'
 import { Box, Flex } from '@chakra-ui/react'
 import { useEffect } from 'react'
 import { useTranslation } from 'react-i18next'
